requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
anyio=3.5.0=py310haa95532_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py310h2bbff1b_0
asttokens=2.0.5=pyhd3eb1b0_0
attrs=21.4.0=pyhd3eb1b0_0
babel=2.9.1=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py310haa95532_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
bottleneck=1.3.5=py310h9128911_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
brotlipy=0.7.0=py310h2bbff1b_1002
bzip2=1.0.8=he774522_0
ca-certificates=2022.07.19=haa95532_0
certifi=2022.9.24=py310haa95532_0
cffi=1.15.1=py310h2bbff1b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.5=py310haa95532_0
cryptography=37.0.1=py310h21b164f_0
cudatoolkit=11.3.1=h59b6b97_2
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.5.1=py310hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.4=py310haa95532_0
executing=0.8.3=pyhd3eb1b0_0
fftw=3.3.9=h2bbff1b_1
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=hd328e21_0
glib=2.69.1=h5dc1a3c_1
gst-plugins-base=1.18.5=h9e645db_0
gstreamer=1.18.5=hd78058f_0
icc_rt=2022.1.0=h6049295_2
icu=58.2=ha925a31_3
idna=3.4=py310haa95532_0
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.15.2=py310haa95532_0
ipython=8.4.0=py310haa95532_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.18.1=py310haa95532_1
jinja2=3.0.3=pyhd3eb1b0_0
joblib=1.1.1=py310haa95532_0
jpeg=9e=h2bbff1b_0
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.16.0=py310haa95532_0
jupyter_client=7.3.5=py310haa95532_0
jupyter_core=4.11.1=py310haa95532_0
jupyter_server=1.18.1=py310haa95532_0
jupyterlab=3.4.4=py310haa95532_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.15.2=py310haa95532_0
kiwisolver=1.4.2=py310hd77b12b_0
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libclang=12.0.0=default_h627e005_2
libdeflate=1.8=h2bbff1b_5
libffi=3.4.2=hd77b12b_4
libiconv=1.16=h2bbff1b_2
libogg=1.3.5=h2bbff1b_1
libpng=1.6.37=h2a8f88b_0
libsodium=1.0.18=h62dcd97_0
libtiff=4.4.0=h8a3f274_0
libuv=1.40.0=he774522_0
libvorbis=1.3.7=he774522_0
libwebp=1.2.4=h2bbff1b_0
libwebp-base=1.2.4=h2bbff1b_0
libxml2=2.9.14=h0ad7f3c_0
libxslt=1.1.35=h2bbff1b_0
lz4-c=1.9.3=h2bbff1b_1
markupsafe=2.1.1=py310h2bbff1b_0
matplotlib=3.5.2=py310haa95532_0
matplotlib-base=3.5.2=py310hd77b12b_0
matplotlib-inline=0.1.6=py310haa95532_0
mistune=0.8.4=py310h2bbff1b_1000
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py310h2bbff1b_0
mkl_fft=1.3.1=py310ha0764ea_0
mkl_random=1.2.2=py310h4ed8f06_0
munkres=1.1.4=py_0
nbclassic=0.3.5=pyhd3eb1b0_0
nbclient=0.5.13=py310haa95532_0
nbconvert=6.4.4=py310haa95532_0
nbformat=5.5.0=py310haa95532_0
nest-asyncio=1.5.5=py310haa95532_0
notebook=6.4.12=py310haa95532_0
numexpr=2.8.3=py310hb57aa6b_0
numpy=1.23.4=pypi_0
numpy-base=1.23.3=py310h04254f7_0
opencv-python=4.6.0.66=pypi_0
openssl=1.1.1q=h2bbff1b_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.4=py310hd77b12b_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=hd77b12b_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=py310hdc2b20a_1
pip=22.2.2=py310haa95532_0
ply=3.11=py310haa95532_0
prometheus_client=0.14.1=py310haa95532_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
psutil=5.9.0=py310h2bbff1b_0
pure_eval=0.2.2=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py310haa95532_0
pyqt=5.15.7=py310hd77b12b_0
pyqt5-sip=12.11.0=py310hd77b12b_0
pyrsistent=0.18.0=py310h2bbff1b_0
pysocks=1.7.1=py310haa95532_0
python=3.10.6=hbb2ffb3_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py310haa95532_0
pytorch=1.12.1=py3.10_cuda11.3_cudnn8_0
pytorch-mutex=1.0=cuda
pytz=2022.1=py310haa95532_0
pywin32=302=py310h2bbff1b_2
pywinpty=2.0.2=py310h5da7b33_0
pyyaml=6.0=pypi_0
pyzmq=23.2.0=py310hd77b12b_0
qt-main=5.15.2=he8e5bd7_7
qt-webengine=5.15.9=hb9a9bb5_4
qtwebkit=5.212=h3ad3cdb_4
requests=2.28.1=py310haa95532_0
scikit-learn=1.1.2=py310hd77b12b_0
scipy=1.9.1=py310h86744a3_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=63.4.1=py310haa95532_0
sip=6.6.2=py310hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.2.0=py310haa95532_1
soupsieve=2.3.2.post1=py310haa95532_0
sqlite=3.39.3=h2bbff1b_0
stack_data=0.2.0=pyhd3eb1b0_0
terminado=0.13.1=py310haa95532_0
testpath=0.6.0=py310haa95532_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pyhd3eb1b0_0
torchmetrics=0.10.0=pypi_0
torchvision=0.13.1=py310_cu113
tornado=6.2=py310h2bbff1b_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=4.3.0=py310haa95532_0
typing_extensions=4.3.0=py310haa95532_0
tzdata=2022e=h04d1e81_0
urllib3=1.26.12=py310haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py310haa95532_1
websocket-client=0.58.0=py310haa95532_4
wheel=0.37.1=pyhd3eb1b0_0
win_inet_pton=1.1.0=py310haa95532_0
wincertstore=0.2=py310haa95532_2
winpty=0.4.3=4
xz=5.2.6=h8cc25b3_0
zeromq=4.3.4=hd77b12b_0
zlib=1.2.12=h8cc25b3_3
zstd=1.5.2=h19a0ad4_0
